clk: tegra: Round up the peak thermal floor voltage
When comparing the peak thermal floor voltage for the CPU between
Linux v3.10 and v4.4 there is a delta of 7mV (957mV versus 950mv).
This is because the peak thermal floor voltage for CPU in Linux
v4.4 is not rounded up where as it is in Linux v3.10. Align Linux
v4.4 with v3.10 by rounding up the peak thermal floor voltage.

Also given that there is both a tegra_dfll_get_thermal_floor()
and tegra_dfll_get_thermal_floor_mv(), update the name of the
tegra_dfll_get_thermal_floor() to be
tegra_dfll_get_peak_thermal_floor_mv() so it is clear what the
difference is.

Bug 1811732

Change-Id: Icc093e88c1421be2ca55325569187e92ab733259
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: http://git-master/r/1284920
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>