dc: ext: add dc open count
Add an atomic field to keep track of how many dc handles
are currently open. This field will be incremented on
open, and decremented on release.

This field can be used to determine if the userspace client
that's interacting with display is still running, and to
trigger a reset if it no longer is.

Bug 1847292
JIRA TDS-1646

Change-Id: I7466abb8449553fd097dbd8d369e8471ba799392
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/1287701
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mitch Luban <mluban@nvidia.com>