ARM64: dts: move to upstream DT clk IDs
The t210-clk-init-table is not used by kernel 4.4, so we can safely only build
this for 3.18. There are however references to i2c_dfll and
dfll_max77621_parms, so we keep those even for 4.4. It's unlikely this will
actually work, but at least it keeps the abca board build happy.
Further work is likely needed to move this board to kernel 4.4.

Bug 1843228

Change-Id: I2441f8302deb5bca64a0138f94c2284106824ab7
Signed-off-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-on: http://git-master/r/1287533
Reviewed-by: Jon Mayo <jmayo@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Timo Alho <talho@nvidia.com>