clk: tegra: Fix compilation when dvfs is not enabled
Commit 92da7e92674a ('clk: tegra: Round up the peak thermal floor
voltage') added a function to return the peak thermal floor voltage
and this function internally makes a call to tegra_round_voltage().
When CONFIG_TEGRA_DVFS is not enabled, but the DFLL driver is enabled
the function tegra_round_voltage() is undefined and breaks the kernel
compilation. Fix this by moving the call to tegra_round_voltage()
outside of the DFLL driver and into the DVFS driver where it is used
and will always be defined.

Add some WARN statements to catch any calls to get the peak thermal
floor voltage where the DFLL driver has not yet been initialised.

Change-Id: If3c891a0654c5e7797aea7ba0cece891ae1ce8f3
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: http://git-master/r/1291535
Tested-by: Harvey Hsieh <hhsieh@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-by: Jon Mayo <jmayo@nvidia.com>