gpio: tegra186: remove kernel version checks
The gpio handler prototype was changed from k318 to k44 and to
support both version of kernel, the code implementation was done
with Linux kernel version check.

Now T186 support is deprecated from before k44 and only supported
in future version of kernel, remove the code implemented for below
k44.

Change-Id: I855c873648b5ecdabf22ac2287bd126e6b894178
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1292194
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>