nvdisp: program and reset default IMP values
This patch contains the following changes:
1) Program default cursor and window ihub settings
on head enable. These default settings will currently
be hardcoded to the power-on values.
2) Add a function to reset the current IMP state back
to its boot-time configuration.

These changes are necessary for usecases where an IMP-aware
client is abruptly killed, and control is handed over to a
client which isn't IMP-aware and has no means of re-configuring
the settings which were left over. Forcing the driver to
explicitly reset the current settings in these circumstances
ensures that any new client that comes up will start from a
blank slate.

Bug 1847292
JIRA TDS-1646

Change-Id: I4c906029d3d552f3aa55bd29a65d28e3a33d5bf5
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/1287705
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>