ARM64: dts correct sclk clock for BPMP
The BPMP driver looks for a clock named sclk rather than sbus. Correct this
and also use the correct DT id. Before this patch, we were just lucky that
there is also a con id sclk which points to the sclk clock. The BPMP driver
only does enable on this clock, so it never really affected sclk.
As this is the last user of the legacy tegra210-clk.h file, we can remove
the include.

Bug 1843228

Change-Id: Iecf6f6bd8bb9942960064d19db4f3c4b96410471
Signed-off-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-on: http://git-master/r/1293265
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>