ARM64: config: use unified PMC driver for T186
Use unified PMC driver which is single driver for
Tegra210 and Tegra186. This will help in kernel unification
and removing of PMC driver for T186 which does the duplication.
The pad control driver for Tegra is converted to generic
pad control driver and hence it is not required to have
SoC specific pad control Get rid of the SoC specific config.

boot.img size decreased by 2048 bytes.

Bug 200256189

Change-Id: I2a00a4a5764fa2966530f4ac3ec1a085847b8cb7
Signed-off-by: Venkat Reddy Talla <vreddytalla@nvidia.com>
Reviewed-on: http://git-master/r/1299839
Tested-by: Timo Alho <talho@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Timo Alho <talho@nvidia.com>