pcie: host: tegra: DW unaligned EP CS access
this patch enables DoubleWord-unaligned end point
config space access for byte,word sizes where it is
relevant. With this, it is possible to access only
specific byte/word during read-modify-write operations
thereby avoiding to perform read-modify-write on all 32-bits
which may accidentally write back '1' to registers
that have the property of write-1-to-clear.

Bug 1866971

Change-Id: Ibc2ff3e9861c05a95733249e4541f700bb2ef096
Signed-off-by: Vidya Sagar <vidyas@nvidia.com>
Reviewed-on: http://git-master/r/1295215
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>