nvdisp: remove IMP node before free in reset func
In the IMP reset function, if the driver detects an error
condition after trying to reserve the max bandwidth settings,
it will jump to the end of the function and free the default
IMP settings struct without removing it from the global list.
This will cause memory corruption problems later on since the
driver is essentially leaving a stale entry in the queue that has
already been freed.

Bug 1847292

Change-Id: Icbdd527270045d06560f803f8d2f575d6bee5605
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/1298897
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Aron Wong <awong@nvidia.com>
Tested-by: Aron Wong <awong@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>