Revert "video: tegra: dc: Fix sw hotplug node"
This reverts commit 94a86456745c058c98ff7b3880579aba138d9a25.
It does not "defer" handling of DC connection status,
rather prevents the status update from happening.
This is because the execution path first disables the IRQ
that carries HPD events.
The previous commit then removed the update to the connection status,
so the status will not change
until the IRQ is manually re-enabled later on.
This leads to system no longer being able to
trigger FPDLink3 detection at run-time.

JIRA: EVLR-950
Bug 200186449

Change-Id: Iece39c5963914b3cc14c9ca85fd1000b4ecaf218
Signed-off-by: Tow Wang <toww@nvidia.com>
Reviewed-on: http://git-master/r/1299033
Reviewed-by: Sungwook Kim <sungwookk@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>