video: tegra: Add scanline tracepoint
Add vblank and vpulse3 scanline tracepoint.

Accurate timestamps of these events are embedded in the
ftrace prints and they are extrapolations based on
current scanline and timestamp (timebase is cycle counter TSC).

It can be enabled/disabled by writing to this debugfs node:
	tracing/events/nvdisp/nvdisp_scanline

The ftrace print includes display controller #, frame # and scanline #
in addition to timestamp in nsec.

The scanline # for VBLANK is specially indicated using value U16_MAX.

Bug 1858948

Change-Id: Ib727bcbeb935f39df0e4b4443c97a9b537cf053c
Signed-off-by: Arun Kannan <akannan@nvidia.com>
Reviewed-on: http://git-master/r/1300684
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>