hdcp: repeater fix for DP
There is a hardware bug in DP which hardware
does not expose an SOR register for reading a
secret value M0. This secret value is needed for
vprime validation for repeater use case. To work-
around this hardware bug, we start the HDCP
authentication as a DP repeater and switch back to
TMDS for vprime verification. Once vprime verification
is completed on TMDS, we switch back to DP receiver and
continue HDCP authentication as a DP receiver.

This patch also fixes the revocation check for receiver
for DP.

Bug 1867887

Change-Id: I103e14d762d3b1b925cfdde5d92aa1f14b545161
Signed-off-by: Pranami Bhattacharya <pranamib@nvidia.com>
(cherry picked from commit 0dec24ffecb54501fee639e66aa7df2bd2e78f6a)
Reviewed-on: http://git-master/r/1305671
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>