video: tegra: host: Disable genpd runtime PM workaround on 4.9
On kernel 4.9, genpd doesn't disable runtime PM for a device
when starting system suspend for it. Therefore we no longer
need the hack to re-enable runtime PM for host1x for suspension.
In fact, this workaround breaks further host1x usage after
resume due to runtime PM staying disabled, so disable the
workaround on kernel 4.9

Bug 1874678

Change-Id: Ia2e02b47534b5c6d735037bfa94df253ae78f4e1
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: http://git-master/r/1306216
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>