nvdisp: check if DC is active for IMP
For IMP, the driver currently checks the "dc->enabled"
flag before programming a given head and waiting for
the state to promote. However, this isn't sufficient
since it doesn't guarantee that the head hasn't been
STOP'ed. This is due to the fact that the display control
mode is set in the "out_ops->enable" callback. For example,
if link training fails during DC enable for a DP head, the
LT driver will detach the SOR and put DC in STOP mode.
In this case, even though the "out_ops->enable" callback
failed, the "dc->enabled" flag will still be set to true.

If DC is in STOP mode, state promotion should occur
instantaneously. However, the FRAME_END/VBLANK worker that's
responsible for specifically checking whether COMMON_ACT_REQ
has cleared and for signaling the COMMON promotion workqueue
won't be scheduled. This change adds a simple inline function
that checks both the "dc->enabled" flag and the display control
mode, and uses this function to determine whether the current
process needs to wait for the workqueue to be signaled.

Bug 200280181

Change-Id: I158164427f8c6b7aab1a7249c839abbada1bf21e
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/1311313
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>