Revert "TEMP: crypto: tegra-se: disable suspend-resume handlers"
SE suspend-resume handlers were reverted earlier as
as it was causing kernel panic during suspend.

The reason for kernel panic is Entropy is being used during
suspend and there are issues when enabling Entorpy clock.
Now on TOT, there are no issues in enabling of Entropy clock.

Reverting this change to enable suspend-resume back.

Bug 200281297

This reverts commit f57cf62ad9bec549bd252c5cc6d7884e74d292c9.

Change-Id: I23fd1b9447613af0840bfbe3deb023d8e121b1c8
Signed-off-by: Mallikarjun Kasoju <mkasoju@nvidia.com>
Reviewed-on: http://git-master/r/1456066
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Signed-off-by: Shravani Dingari <shravanid@nvidia.com>