dc: check ret val of COMMON channel reserve
The implementation of tegra_dc_reserve_common_channel()
is being changed so that it returns an error code if the
given process was not able to successfully wait for the
COMMON channel to be released. As such, the driver needs
to check the return value of tegra_dc_reserve_common_channel()
accordingly.

Bug 200281658

Change-Id: I313df0c91113f7559c283e6f0043c7ecc93c4b7e
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/1310133
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-by: Arun Swain <arswain@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>