nvdisp: use ___wait_event for COMMON channel
Replace the current wq mechanism used for IMP with the
___wait_event() macro. This macro is preferable since it
allows the driver to specify a bounded timeout value while
setting the current task state to TASK_UNINTERRUPTIBLE.

The driver should set the task state to TASK_UNINTERRUPTIBLE
to avoid any spurious wakeups due to signals/traps. In the
current IMP implementation, if a process prematurely wakes up
while waiting for the COMMON channel, there's no way to either
recover or retry. This is due to the fact that all the pending
waiters need to be aligned to the global IMP settings queue,
and should only be woken up one-by-one in the exact order that
they were queued.

Bug 200281658

Change-Id: I029530c8b437db63859cf649e7c8025b8f01c8ab
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/1310132
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Arun Swain <arswain@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>