platform: tegra: rtcpu: vi-notify: schedule notify
The shared mailbox is emptied when the HSP notify function returns. The
mailbox should be emptied before starting to process the IVC
channels. This way we get a new interrupt if new messages are sent to
the channel after old ones have been processed.

Schedule the IVC handling as bottom-half work instead of trying to drain
the IVC channel in the interrupt handler.

Bug 200239989
Bug 1854467
Bug 1883764

Change-Id: I4c30867cc0540231ead442eb50cd9a9a2a12ebee
Reviewed-on: http://git-master/r/1313938
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Tested-by: Pekka Pessi <ppessi@nvidia.com>
Reviewed-by: Ashish Singh <assingh@nvidia.com>
Reviewed-by: Mika Liljeberg <mliljeberg@nvidia.com>
Tested-by: Mika Liljeberg <mliljeberg@nvidia.com>
Reviewed-by: Aki Niemi <aniemi@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Jihoon Bang <jbang@nvidia.com>