video: tegra: host: Fix suspend hack in non-GPD use
The hack to use runtime PM during suspension on kernel 4.4
is only required if Host1x is attached to a power domain,
and if it is not, breaks Host1x after a suspend-resume cycle.

Disable the hack when Host1x is not attached to a power domain.

Jira HOSTX-182
Bug 1882880

Change-Id: I943653b842b22cdf01b84cab7dfd93435f705f78
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: http://git-master/r/1313096
Tested-by: Vignesh Radhakrishnan <vigneshr@nvidia.com>
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Arto Merilainen <amerilainen@nvidia.com>