hdcp: do not allow duplicate id node to be created
The nv/dphdcp structure gets created during the hdmi/dp init
call. For a multiple HDMI/DP use case, we should not create
duplicate nodes for the same head. Say /dev/nvhdcp0 is already
created for head 0, we should not attempt to create the same
node again and return an error. This change handles this case.

This change assumes that if both HDMI and DP displays are
active, different SORs would be powering them.

Bug 1874005

Change-Id: I3a1311094880dcee804d932fec5af9c4dccf3acb
Signed-off-by: Pranami Bhattacharya <pranamib@nvidia.com>
Reviewed-on: http://git-master/r/1313710
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Santosh Galma <galmar@nvidia.com>
Reviewed-by: Pradeepan Patra <ppatra@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>