dc: add power domain hw data
The display power domains for a given chip are fixed.
As such, any information describing them should be added
as part of the hw data for each SOC.

This change adds a generic struct that tracks all relevant fields
for each power domain, as well as APIs that can used to retrieve
those fields. The following information is stored in the struct:
1) Owner head
2) Member heads
3) Member windows
4) Domain clocks
5) Current ref count
6) Powergate partition id

The hw data for each SOC will contain a reference to a table of
these structs. Based on the SOC, the display driver will be
responsible for populating its hw data with the appropriate table.

TDS-1545

Change-Id: I45eb23693cc835d57d716b0b17805168db0b2560
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/1316162
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>