firmware: tegra: do not call msleep() in resume
On VDK, we call msleep() to wait on channels to become ready. This will
cause a problem during suspend->resume if we arrive in native_resume()
before the timer is resumed. If this happens, we never return from the
msleep() call. To fix this, use polling during suspend->resume().

Bug 1880883

Change-Id: I3af53cea356becf0012d92b9b1c3626283b99f68
Signed-off-by: Sivaram Nair <sivaramn@nvidia.com>
Reviewed-on: http://git-master/r/1490589
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Tested-by: Vignesh Radhakrishnan <vigneshr@nvidia.com>
Reviewed-by: Vignesh Radhakrishnan <vigneshr@nvidia.com>
Reviewed-by: Krishna Sitaraman <ksitaraman@nvidia.com>