panel: add label to panel nodes
As per new redesign of the driver, panel selection will be
performed using "nvidia,out-panel" property which references
the phandle of the panel node. In order to reference using
phandle, label is required for the node. Add label for each
panel and their corresponding back-light node as well.

TDS-1544

Change-Id: I74ac078dd48086e2c1914b84ef45d4bca56fabe5
Signed-off-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-on: http://git-master/r/1319138
Reviewed-by: Shu Zhong <shuz@nvidia.com>
GVS: Gerrit_Virtual_Submit