t18x: quill: add "nvidia,active-panel" property
As per the redesign of the driver, active display panel on the given
platform will be selected by property "nvidia,active-panel". Driver
first searches the active panel based on display board-id and if
it doesn't find one, it will look for this property for the
active panel on the given platform.

TDS-1544

Change-Id: I415b84ef1afe854349886642ea3873715c308ed4
Signed-off-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-on: http://git-master/r/1319158
Reviewed-by: Shu Zhong <shuz@nvidia.com>
GVS: Gerrit_Virtual_Submit