video: dc: add support for nsors and dc version
In order to scale the driver for multiple displays, add
number of SORs per SOC and DC version in the HW data. Add
corresponding utility APIs to query number of SORs and DC version
so that rest of the driver can make run-time decisions instead of
using compile time if-defs.

TDS-1543

Change-Id: I77ce67e5e776bb07ac4b4889ecb9693038f8864e
Signed-off-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-on: http://git-master/r/1319525
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>