video: tegra: remove unused tegra-timer.h header
tegra-timer.h header was included as part of a change that added
te_vrr_sec fastcall, which seems unrelated. There doesn't seem
to be any API that is being used from this header. Since this
is not an upstream header file, this is not available in
mainline kernel. Hence removing this header.

bug 1846312

Change-Id: I81dfe5285a404442c052d5313119a19367372c90
Signed-off-by: Naveen Kumar S <nkumars@nvidia.com>
Reviewed-on: http://git-master/r/1321953
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Mikko Perttunen <mperttunen@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-by: Venu Byravarasu <vbyravarasu@nvidia.com>