dp: add proper locking in irq evt worker
If the IRQ event worker detects that there's a loss of
synchronization on any of the DP lanes, it will schedule
a new round of link training. At the start of this process,
the driver detaches the SOR and stops DC, and re-attaches
the SOR and starts DC if training succeeds. Additionally,
the logic that's responsible for attaching the SOR and starting
DC sets the WRITE_MUX to ACTIVE and restores it at the end.

As such, there can be potential race conditions in which the
above programming coincides with other state changes on the
same head. In order to prevent these race conditions, this patch
updates the IRQ event worker so that it grabs the COMMON channel
barrier and dc mutex before scheduling any new link training work,
and so that it releases both only after it detects that the scheduled
work has completed, or timed out.

Bug 200280181

Change-Id: I46698cd6d990597beae8df288cc3b5f94868ed60
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/1328094
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>