video: tegra: pva: BlockLiner support
Add block_height_log2 variable to the surface
descriptor to pass the block height from
user space.

To be backward compatible, layout and
block_height_log2 are defined as
16bit variables.

Set bit 39 for block linear surfaces in the address field.
This bit is used for indicating that memory subsystem
should convert the block linear format into common
block linear format that is used by other engines in Tegra.
The bit in itself is dropped before making the address
translation in SMMU.

JIRA PVA-93

Change-Id: I9b482412dfcbe845951d01e62abd20ff8d53d249
Signed-off-by: Vinod G <vinodg@nvidia.com>
Reviewed-on: http://git-master/r/1327181
Reviewed-by: Arto Merilainen <amerilainen@nvidia.com>
Tested-by: Arto Merilainen <amerilainen@nvidia.com>