video: tegra: host: Use VM threshold registers
Host1x implements two different syncpoint threshold register ranges
for setting the interrupts:
* VM specific thresholds which trigger VM threshold interrupt
* Common threshold ranges which are routed to CPU interrupts

Currently nvhost uses the common threshold registers but it still
uses the VM specific syncpoint interrupts. Currently this works
due to a model bug, however, it will not be functional in newer
simulator models or silicon.

This change updates code to conditionally use the VM specific
threshold registers: When host1x is being turned on, the code
validates if the old threshold register is functional. If yes,
the code keeps using it. If not, the code moves to use the new
threshold range.

JIRA HOSTX-192
Bug 200293149

Change-Id: Ifd3a4f23c3212826e20268725f182562d9e2d793
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/1325300
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mikko Perttunen <mperttunen@nvidia.com>