gpu: nvgpu: debugging API update
Update the debugging APIs in nvgpu to be more coherent and less dependent
on Linux kernel APIs and data structures. The old API is maintained for
backwards compatibility but the prints themselves are changed.

The old API now prints stuff like this:

  [   46.376077] nvgpu:              gk20a_channel_suspend:3430 [DBG]  suspend channel 508
  [   46.376242] nvgpu:              gk20a_channel_suspend:3430 [DBG]  suspend channel 509
  [   46.376359] nvgpu:              gk20a_channel_suspend:3430 [DBG]  suspend channel 510
  [   46.376466] nvgpu:              gk20a_channel_suspend:3430 [DBG]  suspend channel 511
  [   46.376604] nvgpu:   gk20a_fifo_update_runlist_locked:3090 [DBG]  runlist_id : 0, switch to new buffer 0x       4003a0000
  [   46.378348] nvgpu:                  gk20a_mm_l2_flush:5259 [DBG]  l2_flush_dirty 0x2
  [   46.378562] nvgpu:                     clk_config_pll:170  [DBG]  low_PL 1(div1), high_PL 2(div2)
  [   46.378684] nvgpu:                     clk_config_pll:170  [DBG]  low_PL 8(div8), high_PL 17(div17)

Each line starts with 'nvgpu:' since this is the nvgpu driver. Then there's
the function name which is is right justified and post-fixed with a line
number. This keeps all statements lined up so long as the function name does
not exceed the length of characters alloted for function names (33 currently).
Lines are also left justified with 4 available spaces since there are currently
no source files with over 9999 lines. The type of message (DBG, WRN, ERR) is
then printed in brackets. Finally there's whatever message is to be printed.

The new API will be largely the same except there will be an additional
print of the GPU that the message has originated from in order to help debug
cases where both an iGPU and dGPU are active. Also the implicit newline
added in the legacy API has been removed. This is inline with essentially
every other print function in the C world. There are numerous places where
'\n' is added despite it not being needed which results in log files being
littered with blank lines. This makes the logs much harder to read.

Bug ...

Change-Id: I7bc87f2dbbaebf6eb6c9de1a629a0e2963e1804c
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/1329487
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>