dc: fix WARN_ON for boot_out_type size
boot_out_type is a static array that caches the previous
output type for each DC during an outtype switch. This array
is sized according to the max number of heads that can be
supported across all relevant SOCs. As such, update the WARN_ON
so that it only flags scenarios in which boot_out_type is too small,
instead of checking for strict equality.

Bug 1896053

Change-Id: I17d1a113ada84cd174135eedc526417191d3e4fa
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: http://git-master/r/1329732
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>