ASoC: tegra-alt: remove dspk config check
Card registration fails with LKM changes, as the dependency between
of this driver on dspk driver is not explicitly mentioned. Current
patch removes the config check CONFIG_SND_SOC_TEGRA186_DSPK_ALT.
This is anyway not needed as correct CIF/DAPs are chosen based on
the if/else checks we have.

Bug 200291328

Change-Id: If213453114a7a41d5c655f26da0da0be366825d0
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: http://git-master/r/1329851
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ravindra Lokhande <rlokhande@nvidia.com>