video: tegra: dc: ext: workq becomes kthread
Replace the usage of workq inside the dc flipworker with a
kthread.  This allows us explicit scheduling control over
the kthread running this work, including the ability to make
that thread realtime.  This allows us to tighten the latency
and execution time of the dc flip work.  Since the the display
is updated when the dc work completes instead of on a
vsync pulse in vrr mode the latency and scheduling of the
work becomes much more important.

Bug 1657411

Change-Id: Ib8d5cec0bc9c67f788e5efee81aa82a465d00842
Signed-off-by: David Lock <dlock@nvidia.com>
Reviewed-on: http://git-master/r/1330185
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>