ARM64: dts: tegra210: common: emc-cpu table revise
K4.4 uses A57 frequencies in the entire DVFS range which causes
the CPU frequency points are slightly difference with that in
K3.10. The impact is that the power consumption would be higher
when CPU is running around 700MHz because in K3.10 cpu@710MHz with
emc@408MHz while in K4.4 cpu@714MHz with emc@665MHz based on the
current emc-cpu table.
This patch is to set emc@408MHz for cpu@714MHz as similar in K3.10.

Bug 200282764
Signed-off-by: Aaron Huang <aaronh@nvidia.com>

Change-Id: I5f0d2609e1f21e664e9ab367253894fce0f0aed1
Reviewed-on: http://git-master/r/1458926
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Aaron Huang <aaronh@nvidia.com>
Tested-by: Aaron Huang <aaronh@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Wen Yi <wyi@nvidia.com>
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>